PK-8
Fix in "Kontakt" tab
